+++ iblrig/gui/wizard.py
@@ -11,11 +11,9 @@ import sys
 import traceback
 import webbrowser
 from collections import OrderedDict
-from collections.abc import Iterable
 from dataclasses import dataclass
 from pathlib import Path
 
-import numpy as np
 import pyqtgraph as pg
 from pydantic import ValidationError
 from PyQt5 import QtCore, QtGui, QtWidgets